[spec] Make bikeshed check references to Private Aggregation API
Currently, all references to the private-aggregation-api spec are
declared in the "anchors" block. According to the bikeshed docs, this is
how you link to dfns in specs that aren't part in the autolinking
database[1]. As a result, bikeshed hasn't actually been checking our
references, nor has it always been generating the correct links[2].

Private Aggregation spec is already in the database[3], so I think we
can promote these to actual references.

This commit makes the necessary adjustments to the "link-defaults"
block. With this change, invalid references to the Private Aggregation
spec are now compilation errors.

[1]: https://speced.github.io/bikeshed/#custom-dfns
[2]: For instance, clicking "context ID" in step 10 of "obtain the
     pre-specified report parameters" links to a non-existent anchor on
     the Private Aggregation spec.
[3]: https://github.com/tobie/specref/blob/3569d1ecfa298461f3de0586ed2d13ffb000fa29/refs/browser-specs.json#L1179